factor out help button into a component
